Changed the database connection to Postgres and wrote a guide on how to set up a local connection
