+++ eclipselink/jpa-basic/pom.xml
@@ -46,25 +46,13 @@
       <scope>test</scope>
       </dependency>
 
-      <!-- https://mvnrepository.com/artifact/org.eclipse.persistence/eclipselink -->
+   
+
       <dependency>
-          <groupId>org.eclipse.persistence</groupId>
-          <artifactId>eclipselink</artifactId>
-          <version>2.7.7</version>
+          <groupId>org.postgresql</groupId>
+          <artifactId>postgresql</artifactId>
+          <scope>runtime</scope>
       </dependency>
-
-      <!-- https://mvnrepository.com/artifact/org.apache.derby/derbyclient -->
-     <dependency>
-        <groupId>org.apache.derby</groupId>
-        <artifactId>derby</artifactId>
-        <version>10.15.2.0</version>
-    </dependency>
-
-    <dependency>
-        <groupId>org.apache.derby</groupId>
-        <artifactId>derbytools</artifactId>
-        <version>10.15.2.0</version>
-    </dependency>
 
       <dependency>
           <groupId>org.springframework.boot</groupId>
