+++ eclipselink/jpa-basic/src/main/java/no/hvl/dat250/jpa/basicexample/entities/Poll.java
@@ -50,7 +50,7 @@ public class Poll {
     }
 
     public void removeCreator(){
-        assert(this.creator != null);
+        if(this.creator == null)return;
         this.creator.getCreatedPolls().remove(this);
         setCreator(null);
     }
