--- eclipselink/jpa-basic/src/main/resources/META-INF/persistence.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<persistence xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd"
-             version="2.0" xmlns="http://java.sun.com/xml/ns/persistence">
-    <persistence-unit name="votingsystem" transaction-type="RESOURCE_LOCAL">
-        <class>no.hvl.dat250.jpa.basicexample.entities.UserClass</class>
-        <class>no.hvl.dat250.jpa.basicexample.entities.Poll</class>
-        <class>no.hvl.dat250.jpa.basicexample.entities.Vote</class>
-
-        <properties>
-            <property name="javax.persistence.jdbc.driver" value="org.apache.derby.jdbc.EmbeddedDriver"/>
-            <property name="javax.persistence.jdbc.url"
-                      value="jdbc:derby:$PROJECT_DIR$\testdb;create=true"/>
-            <property name="javax.persistence.jdbc.user" value="test"/>
-            <property name="javax.persistence.jdbc.password" value="test"/>
-
-            <property name="eclipselink.ddl-generation" value="drop-and-create-tables"/>
-            <property name="eclipselink.ddl-generation.output-mode" value="database"/>
-        </properties>
-
-    </persistence-unit>
-</persistence>
