+++ eclipselink/jpa-basic/src/main/resources/application.properties
@@ -1,0 +1,12 @@
+spring.jpa.database=POSTGRESQL
+spring.sql.init.platform=postgres
+spring.datasource.url=${DATABASE_URL}
+spring.datasource.username=${DATABASE_USERNAME}
+spring.datasource.password=${DATABASE_PASSWORD}
+!DATABASE_URL=jdbc:postgresql://localhost:5432/feedapp;DATABASE_USERNAME=feedapp;DATABASE_PASSWORD=feedapp
+spring.jpa.show-sql=true
+spring.jpa.generate-ddl=true
+!TODO This should be changed from create-drop when in production
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
+
